requirements.txt
aiofiles==0.7.0
asgiref==3.4.1
bcrypt==3.2.0
certifi==2021.10.8
cffi==1.15.0
click==8.0.1
colorama==0.4.4
cryptography==36.0.0
dnspython==2.1.0
email-validator==1.1.3
fastapi==0.68.1
h11==0.12.0
idna==3.3
mysql-connector-python==8.0.27
peewee==3.14.8
protobuf==3.19.1
pycparser==2.21
pydantic==1.8.2
PyJWT==2.3.0
PyMySQL==1.0.2
python-dotenv==0.19.2
python-http-client==3.3.4
sendgrid==6.9.3
sentry-asgi==0.2.0
sentry-sdk==1.5.0
six==1.16.0
starkbank-ecdsa==2.0.3
starlette==0.14.2
typing-extensions==3.10.0.2
urllib3==1.26.7
uvicorn==0.15.0
